requirements.txt
python-dotenv==1.1.0
Requests==2.32.3
BeautifulSoup4==4.13.3
Selenium==4.30.0
Webdriver-Manager==4.0.2